requirements.txt
PyYAML==3.12
requests==2.18.4
Jinja2==2.10
